+++ at_end2end_test/test/update_verb_test.dart
@@ -1,6 +1,8 @@
+import 'dart:convert';
 import 'dart:math';
 
 import 'package:test/test.dart';
+import 'package:version/version.dart';
 
 import 'commons.dart';
 import 'e2e_test_utils.dart' as e2e;
@@ -67,8 +69,18 @@ void main() {
     ///LLOOKUP VERB - with out @sign does not return value.
     await sh1.writeCommand('llookup:country$atSign_1');
     response = await sh1.read();
-    print('llookup verb response without private key in llookup verb: $response');
-    expect(response, contains('error:AT0015-key not found : country$atSign_1 does not exist in keystore'));
+    print(
+        'llookup verb response without private key in llookup verb: $response');
+    //TODO: remove else block in the next release -  when all servers json encode the error responses
+    var serverResponse = Version.parse(await sh1.getVersion());
+    if (serverResponse > Version(3, 0, 23)) {
+      response = response.replaceFirst('error:', '');
+      var errorMap = jsonDecode(response);
+      expect(errorMap['errorCode'], 'AT0015');
+      assert(errorMap['errorDescription'].contains('key not found'));
+    } else {
+      expect(response, contains('country$atSign_1 does not exist in keystore'));
+    }
   });
 
   test('update verb with special characters', () async {
@@ -76,7 +88,8 @@ void main() {
     await sh1.writeCommand('update:public:passcode$atSign_1 @!ice^&##');
     var response = await sh1.read();
     print('update verb response : $response');
-    assert((!response.contains('Invalid syntax')) && (!response.contains('null')));
+    assert(
+        (!response.contains('Invalid syntax')) && (!response.contains('null')));
 
     ///LLOOKUP VERB
     await sh1.writeCommand('llookup:public:passcode$atSign_1');
@@ -90,7 +103,8 @@ void main() {
     await sh1.writeCommand('update:public:unicode$atSign_1 U+0026');
     var response = await sh1.read();
     print('update verb response : $response');
-    assert((!response.contains('Invalid syntax')) && (!response.contains('null')));
+    assert(
+        (!response.contains('Invalid syntax')) && (!response.contains('null')));
 
     ///LLOOKUP VERB
     await sh1.writeCommand('llookup:public:unicode$atSign_1');
@@ -101,10 +115,12 @@ void main() {
 
   test('update verb with spaces ', () async {
     ///UPDATE VERB
-    await sh1.writeCommand('update:public:message$atSign_1 Hey Hello! welcome to the tests');
+    await sh1.writeCommand(
+        'update:public:message$atSign_1 Hey Hello! welcome to the tests');
     var response = await sh1.read();
     print('update verb response : $response');
-    assert((!response.contains('Invalid syntax')) && (!response.contains('null')));
+    assert(
+        (!response.contains('Invalid syntax')) && (!response.contains('null')));
 
     ///LLOOKUP VERB
     await sh1.writeCommand('llookup:public:message$atSign_1');
@@ -113,12 +129,15 @@ void main() {
     expect(response, contains('data:Hey Hello! welcome to the tests'));
   });
 
-  test('updating same key with different values and doing a llookup ', () async {
+  test('updating same key with different values and doing a llookup ',
+      () async {
     ///UPDATE VERB
-    await sh1.writeCommand('update:public:message$atSign_1 Hey Hello! welcome to the tests');
+    await sh1.writeCommand(
+        'update:public:message$atSign_1 Hey Hello! welcome to the tests');
     var response = await sh1.read();
     print('update verb response : $response');
-    assert((!response.contains('Invalid syntax')) && (!response.contains('null')));
+    assert(
+        (!response.contains('Invalid syntax')) && (!response.contains('null')));
 
     ///LLOOKUP VERB
     await sh1.writeCommand('llookup:public:message$atSign_1');
@@ -126,10 +145,12 @@ void main() {
     print('llookup verb response : $response');
     expect(response, contains('data:Hey Hello! welcome to the tests'));
 
-    await sh1.writeCommand('update:public:message$atSign_1 Hope you are doing good');
+    await sh1
+        .writeCommand('update:public:message$atSign_1 Hope you are doing good');
     response = await sh1.read();
     print('update verb response : $response');
-    assert((!response.contains('Invalid syntax')) && (!response.contains('null')));
+    assert(
+        (!response.contains('Invalid syntax')) && (!response.contains('null')));
 
     ///LLOOKUP VERB
     await sh1.writeCommand('llookup:public:message$atSign_1');
@@ -185,7 +206,8 @@ void main() {
   test('update verb by sharing a cached key ', () async {
     ///UPDATE VERB
     var value = 'joey$lastValue$lastValue';
-    await sh1.writeCommand('update:ttr:-1:$atSign_2:youtube_id$atSign_1 $value');
+    await sh1
+        .writeCommand('update:ttr:-1:$atSign_2:youtube_id$atSign_1 $value');
     var response = await sh1.read();
     print('update verb response : $response');
     assert(
@@ -252,7 +274,7 @@ void main() {
     response = await sh1.read();
     print('llookup verb response : $response');
     expect(response, contains('data:$value'));
-    });
+  });
 
   test('update-llookup for ttl ', () async {
     ///UPDATE VERB
@@ -350,4 +372,18 @@ void main() {
     print('llookup verb response after 4 seconds : $response');
     expect(response, contains('data:null'));
   });
+
+  // commenting this test till next version of server is released
+  // test('update verb key with punctuation - check invalid key ', () async {
+  //   ///UPDATE VERB
+  //   await sh1.writeCommand('update:public:country,current$atSign_1 USA');
+  //   var response = await sh1.read();
+  //   print('update verb response : $response');
+  //   expect(response, startsWith('error:AT0016'));
+  //   expect(response, contains('Invalid key'));
+  //
+  //   // Going to reconnect, because invalid syntax causes server to close connection
+  //   sh1.close();
+  //   sh1 = await e2e.getSocketHandler(atSign_1);
+  // });
 }