+++ at_secondary/at_persistence_secondary_server/lib/src/utils/at_metadata_adapter.dart
@@ -1,6 +1,7 @@
 import 'package:at_commons/at_commons.dart';
 import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
 
+// ignore: non_constant_identifier_names
 AtMetaData? AtMetadataAdapter(Metadata metadata) {
   var atMetadata = AtMetaData();
   atMetadata
@@ -12,7 +13,23 @@ AtMetaData? AtMetadataAdapter(Metadata metadata) {
     ..isEncrypted = metadata.isEncrypted
     ..dataSignature = metadata.dataSignature
     ..sharedKeyEnc = metadata.sharedKeyEnc
-    ..pubKeyCS = metadata.pubKeyCS;
+    ..pubKeyCS = metadata.pubKeyCS
+    ..encoding = metadata.encoding;
 
   return AtMetadataBuilder(newAtMetaData: atMetadata).build();
+}
+
+Metadata metadataAdapter(AtMetaData atMetaData) {
+  var metadata = Metadata()
+    ..ttl = atMetaData.ttl
+    ..ttb = atMetaData.ttb
+    ..ttr = atMetaData.ttr
+    ..ccd = atMetaData.isCascade
+    ..isBinary = atMetaData.isBinary
+    ..isEncrypted = atMetaData.isEncrypted
+    ..dataSignature = atMetaData.dataSignature
+    ..sharedKeyEnc = atMetaData.sharedKeyEnc
+    ..pubKeyCS = atMetaData.pubKeyCS
+    ..encoding = atMetaData.encoding;
+  return metadata;
 }