+++ at_secondary/at_secondary_server/lib/src/connection/base_connection.dart
@@ -6,12 +6,14 @@ import 'package:at_utils/at_logger.dart';
 
 /// Base class for common socket operations
 abstract class BaseConnection extends AtConnection {
-  final Socket? _socket;
+  late final Socket? _socket;
   late AtConnectionMetaData metaData;
   late var logger;
 
-  BaseConnection(this._socket) {
+  BaseConnection(Socket? socket) {
     logger = AtSignLogger(runtimeType.toString());
+    socket?.setOption(SocketOption.tcpNoDelay, true);
+    _socket = socket;
   }
 
   @override
@@ -52,6 +54,7 @@ abstract class BaseConnection extends AtConnection {
       getMetaData().lastAccessed = DateTime.now().toUtc();
     } on Exception catch (e) {
       getMetaData().isStale = true;
+      logger.severe(e.toString());
       throw AtIOException(e.toString());
     }
   }