+++ at_secondary/at_secondary_server/lib/src/exception/global_exception_handler.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/args.dart';
@@ -6,6 +7,7 @@ import 'package:at_persistence_spec/at_persistence_spec.dart';
 import 'package:at_secondary/src/server/at_secondary_impl.dart';
 import 'package:at_server_spec/at_server_spec.dart';
 import 'package:at_utils/at_logger.dart';
+import 'package:version/version.dart';
 
 /// GlobalExceptionHandler class is used to handle all the exceptions in the system.
 var logger = AtSignLogger('GlobalExceptionHandler');
@@ -34,9 +36,9 @@ class GlobalExceptionHandler {
       await _sendResponseForException(exception, atConnection);
       // TODO but do they necessarily need the connection to be closed?
       _closeConnection(atConnection);
-
     } else if (exception is BlockedConnectionException ||
-        exception is InvalidSyntaxException) {
+        exception is InvalidSyntaxException ||
+        exception is InvalidAtKeyException) {
       // This is normal behaviour, log as INFO
       logger.info(exception.toString());
       await _sendResponseForException(exception, atConnection);
@@ -45,13 +47,11 @@ class GlobalExceptionHandler {
       //   BlockedConnectionException thrown when the "from" atsign is on the "do not allow" list
       //   InvalidSyntaxException is thrown because invalid syntax is rude, so we're rude in return
       _closeConnection(atConnection);
-
     } else if (exception is DataStoreException) {
       logger.severe(exception.toString());
       // TODO should we keep the connection open rather than closing it?
       await _sendResponseForException(exception, atConnection);
       _closeConnection(atConnection);
-
     } else if (exception is InboundConnectionLimitException) {
       // This is SEVERE and requires different handling so we use _handleInboundLimit
       logger.severe(exception.toString());
@@ -71,20 +71,9 @@ class GlobalExceptionHandler {
       // TODO Not sure some of these are really worthy of WARNINGS, but let's leave as is for now
       logger.warning(exception.toString());
       await _sendResponseForException(exception, atConnection);
-
-    } else if (exception is ArgParserException) {
-      // TODO [gkc] I suspect this code block is not reachable. Verify and delete
-      logger.shout("Terminating secondary due to ${exception.toString()}");
-      try {
-        await AtSecondaryServerImpl.getInstance().stop();
-      } finally {
-        exit(1);
-      }
-
     } else if (exception is InternalServerError) {
       logger.severe(exception.toString());
       await _handleInternalException(exception, atConnection);
-
     } else {
       logger.shout("Unexpected exception '${exception.toString()}'");
       await _handleInternalException(
@@ -124,11 +113,21 @@ class GlobalExceptionHandler {
         var errorCode = getErrorCode(exception);
         errorCode ??= 'AT0011';
 
-        var errorDescription;
-        if (exception is AtException) {
-          errorDescription = '${getErrorDescription(errorCode)} : ${exception.message}';
+        String errorDescription;
+        // To avoid duplication of error description in the error message add errorDescription only on
+        // the sender side (not on the receiver side).
+        // For example if @alice performs lookup verb to @bob and @bob returns key not found exception
+        // add error description only on @alice (not on @bob)
+        // When connecting to other secondaries for lookup verb, a pol authenticated connection
+        // is established. atConnection.getMetaData().isPolAuthenticated is set to true.
+        // When a user connects to his own secondary, an authenticate connection is created.
+        // atConnection.getMetaData().isAuthenticated is set to true.
+        if (exception is AtException &&
+            atConnection.getMetaData().isAuthenticated) {
+          errorDescription =
+              '${getErrorDescription(errorCode)} : ${exception.message}';
         } else {
-          errorDescription = '${getErrorDescription(errorCode)} : ${exception.toString()}';
+          errorDescription = exception.toString();
         }
         _writeToSocket(atConnection, prompt, errorCode, errorDescription);
       }
@@ -152,6 +151,24 @@ class GlobalExceptionHandler {
 
   void _writeToSocket(AtConnection atConnection, String prompt,
       String? errorCode, String errorDescription) {
+    if (atConnection.getMetaData().clientVersion ==
+        AtConnectionMetaData.clientVersionNotAvailable) {
+      atConnection.write('error:$errorCode-$errorDescription\n$prompt');
+      return;
+    }
+    // The JSON encoding of error message is supported by the client versions greater than 3.0.37
+    if (Version.parse(atConnection.getMetaData().clientVersion) >
+        Version(3, 0, 37)) {
+      logger.info(
+          'Client version supports json encoding.. returning Json encoded error message');
+      var errorJsonMap = {
+        'errorCode': errorCode,
+        'errorDescription': errorDescription
+      };
+      atConnection.write('error:${jsonEncode(errorJsonMap)}\n$prompt');
+      return;
+    }
+    // Defaults to return the error message in string format if all the conditions fails
     atConnection.write('error:$errorCode-$errorDescription\n$prompt');
   }
 }