+++ at_secondary/at_persistence_secondary_server/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 3.0.36
+- fix: skip commit id and sync for signing keys
+- fix: dart analyzer issues
+- chore: upgrade third party dependencies
+## 3.0.35
+- fix: Randomize the cron job's start interval
+- fix: Reduce the default notification expiry duration
+## 3.0.34
+* fix: Reverted dependency on 'meta' package to ^1.7.0 as flutter_test package (currently) requires 1.7.0
+## 3.0.33
+- feat: added key validation to keystore put and create methods
+- chore: upgraded at_commons version to 3.0.24
+## 3.0.32
+- Add 'encoding' to AtMetadata which represents the type of encoding
 ## 3.0.31
 - Invalidate commit log cache on removing entry from commit log
 ## 3.0.30